insights: fix flaky insights test
This test is flaky because the async tagging of the
TransactionId to the insight sometimes takes a long
time to complete. This issue has been seen in other
branches and the code has been removed.

To fix, the code which is testing the value
of waitingTxnFingerprintID has been removed.

For reference, the PR in merged into master is #126524

Resolves: #130348
Release note: None