roachtest: EveryN util without CRDB log
Previously, tests made use of the EveryN utility in "pkg/util/log" that also
interacts with the CRDB log to check verbosity levels. This feature is not used
in `roachtest` as generally CRDB log should not be used explicitly, nor will it
be configured for verbosity levels.

This change adds the utility to `roachtest`, which still leverages the base
utility EveryN from "pgk/util", but without interacting with CRDB log.

Epic: None
Release note: None