sql/catalog: fix panic inside sequence expression parsing
Previously, if the correct overloads were not found for sequence
builtins it was possible for the server to panic. This could happen when
rewriting a CREATE TABLE expression with an invalid sequence builtin
call. To address this, this patch updates the sequence logic to return
the error instead of panicking on it.

Fixes: #133399

Release note (bug fix): Address a panic inside CREATE TABLE AS if
sequence builtin expressions had invalid function overloads.