opt: optimize table func deps for non-unique indexes
When a table is referenced in a query, we build a func-dep set describing
the relationships between the table columns. Among other things, we use
the index key columns for this purpose. There is no benefit to considering
non-unique indexes here because the key columns are always a superset of
the primary key columns. Therefore, this commit skips non-unique indexes to
avoid the extra work. This results in a performance improvement for queries
against tables with many non-unique indexes and many columns.

Epic: None

Release note: None