kvserver: enable leader leases for StoreRangeMergeAbandonedFollowersAutomaticallyGarbageCollected
This commit enables leader leases to run with
TestStoreRangeMergeAbandonedFollowersAutomaticallyGarbageCollected. The
test used to move the lease and stop Raft traffic without waiting for
leadership to follow it. This could cause flakes with leader leases as
if the replica doesn't know who the leader is, it might not be able to
determine the lease status.

References: #136806

Release Note: None