Create test1
sql stats flush could stuck waiting on the waitgroup used so synchronize the flushing of statement and
transaction statistics. In the case where a flush is started and a node is shutdown, it is possible for one of the async tasks to never start, causing the flush to wait forever.

To fix, wait.Done() is called when the either go routine flushing errors. This was the logic that existed prior to sql stats flushed being refactored (See cockroachdb#140673).

Resolves: cockroachdb#142074
Epic: None
Release note: sql stats flush could stuck waiting on the waitgroup used so synchronize the flushing of statement and
transaction statistics. In the case where a flush is started and a node is shutdown, it is possible for one of the async tasks to never start, causing the flush to wait forever.